winex11.drv: Limit fullscreen visible rects to the virtual screen rect.
Fix a regression from 31fda1f4, which allowed the visible rect to be larger than the monitor rect.
After 31fda1f4, CHRONO TRIGGER (613830) sets a window rect slightly larger than the monitor rect and
will change the window rect to the rect it previously set if the game detects a different window rect.
Adding __NET_WM_STATE_FULLSCREEN will cause WMs to move the window to cover exactly the monitor rect.
So the window rect will be repeatedly changed by the WM and the game, causing a flickering effect. Limit
fullscreen visible rects to the virtual screen rect so that the visible rects in winex11.drv are of
the same size as the monitor rect. Thus, adding __NET_WM_STATE_FULLSCREEN won't trigger a size change.

CW-Bug-Id: #25747