fixup! winex11.drv: Ignore fullscreen window config changes.
xinerama_get_fullscreen_monitors() was supposed to be called with fullscreen rects. Thus, we could
assume the monitor indices are all zero when there is only one monitor as an optimization. However,
after baf65867, xinerama_get_fullscreen_monitors() can be called with non-fullscreen rects. So we
need to remove the optimization so that it works correctly for non-fullscreen rects as well.

CW-Bug-Id: #25747