winex11.drv: Update the gl_drawable of the top-level window when the visibility of a child window changes.
needs_client_window_clipping() returns FALSE for a top-level window if its child windows are
invisible. After a child window becomes visible, we should update the gl_drawable of the top-level
window. Otherwise, the gl_drawable might still be rendering on-screen and causes black screen or
flickering effects because the child window is not composited on top.

CW-Bug-Id: #26177